+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## v0.8 - Personalization
+
+### v0.8.0 (July 27, 2020)
+[`Add user authentication`](https://github.com/WillFlame14/jspuyo/pull/47)
+**New features:**
+- User authentication added to facilitate future storage of user data
+  - Sign-in options are currently limited to email/password and Google sign-in
+  - Playing as a Guest is also an option, though data will not be saved between browsers
+  - All locations should now use your username
+  - You can log out under the Profile panel
+- Pages for About, Privacy Policy, and Terms of Service (tbd)
+- 3 new backgrounds!
+- Support for 'FlatColour' skin
+
+**Other changes:**
+- Organized entire filesystem based on production files vs. development files.
+- Very small tuning improvements
+- Fixed a ton of room-related bugs
+
 ## v0.7 - Better rooms, Spectate feature
 
 ### v0.7.3 (July 3, 2020)